Memoize numberFormatter and pluralRules in translate utility
